pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "standards"
version = "2.0.0"
description = "Comprehensive software development standards with Claude-Flow orchestration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "William Zujkowski"}
]
keywords = [
    "standards",
    "documentation",
    "claude",
    "ai",
    "development",
    "best-practices",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]

# Core dependencies (from requirements.txt)
dependencies = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-git-revision-date-localized-plugin>=1.5.0",
    "pymdown-extensions>=10.19.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.14.0",
    "pre-commit>=3.6.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-git-revision-date-localized-plugin>=1.5.0",
    "pymdown-extensions>=10.19.1",
]

[project.urls]
Homepage = "https://github.com/williamzujkowski/standards"
Documentation = "https://williamzujkowski.github.io/standards/"
Repository = "https://github.com/williamzujkowski/standards"
Issues = "https://github.com/williamzujkowski/standards/issues"

# ============================================================================
# Ruff Configuration - Replaces Black, isort, and provides linting
# ============================================================================

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "venv",
    "build",
    "dist",
    "node_modules",
    "__pycache__",
    "*.egg-info",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "archive",
    "vendor",
]

[tool.ruff.format]
# Use double quotes for strings (consistent with black's default)
quote-style = "double"
# Use spaces for indentation
indent-style = "space"
# Respect magic trailing commas
skip-magic-trailing-comma = false
# Auto-detect line endings
line-ending = "auto"
# Format docstrings
docstring-code-format = true
docstring-code-line-length = 80

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "B",      # flake8-bugbear
    "S",      # flake8-bandit (security)
    "I",      # isort (import sorting)
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PL",     # pylint
    "PERF",   # perflint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",    # line too long - handled by formatter
    "E722",    # bare except - sometimes needed
    "S101",    # use of assert detected - allowed in tests
    "S103",    # chmod permissions
    "S105",    # possible hardcoded password - false positives in examples
    "S106",    # possible hardcoded password - false positives in examples
    "S110",    # try-except-pass - acceptable in some cases
    "S112",    # try-except-continue - acceptable in monitoring
    "S324",    # MD5 hash - used for checksums, not security
    "S602",    # subprocess with shell=True - needed in scripts
    "S603",    # subprocess untrusted input - needed in scripts
    "S607",    # partial executable path - needed in scripts
    "B007",    # unused loop variable
    "N802",    # function name should be lowercase - needed for HTTP handlers
    "PLR0913", # too many arguments - sometimes needed
    "PLR2004", # magic value comparison - acceptable for constants
    "ARG001",  # unused function argument - sometimes needed for API compatibility
    "ARG002",  # unused method argument - sometimes needed for API compatibility
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR2004", "ARG"]  # Allow assert and magic values in tests
"scripts/*" = ["S602", "S603", "S607", "T201"]  # Allow subprocess and print in scripts
"monitoring/*" = ["S602", "S607", "E722"]  # Allow shell commands and bare except
"examples/*" = ["S105", "S106"]  # Allow example credentials in examples

[tool.ruff.lint.isort]
# Match the behavior of the old isort configuration
force-single-line = false
lines-after-imports = 2
known-first-party = ["standards"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

# ============================================================================
# Pytest Configuration
# ============================================================================

[tool.pytest.ini_options]
markers = [
    "quality_gate: marks tests as quality gate tests (deselect with '-m \"not quality_gate\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra -q --strict-markers"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# ============================================================================
# Coverage Configuration
# ============================================================================

[tool.coverage.run]
source = ["scripts", "tests"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
